requirements.txt
django>=4
celery
djangorestframework
redis
requests
ffmpeg
pika
asyncio
PyYAML==6.0
coreapi==2.3.3